+++ doc/html/search/groups_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['turbojpeg_281',['TurboJPEG',['../group___turbo_j_p_e_g.html',1,'']]]
+  ['turbojpeg_282',['TurboJPEG',['../group___turbo_j_p_e_g.html',1,'']]]
 ];